Rename command CLI FAC to FASPI (#1)
* Update Cargo.toml

* docs: update Cargo.toml description and exclude files

Update the project description to better reflect its purpose as a CLI generator for FastAPI projects with hexagonal architecture. Add exclude list to prevent unnecessary files from being included in the package

* refactor: rename project from 'fac' to 'faspi'

Update project name, CLI commands, and related documentation to reflect the new name 'faspi'. This change includes modifications to the Cargo.toml, CLI parser, CHANGELOG, and README files, as well as the addition of a new logo asset.

* refactor(cli/parser): remove redundant CLI help text

The CLI help text was duplicated in the code, making it harder to maintain. This commit removes the redundant section to improve code readability and maintainability.